renaming filters-options to filters-content
