+++ visualizer/style.css
@@ -12,7 +12,7 @@
 @import "./info-box.css";
 @import "@nearform/clinic-common/spinner/style.css";
 @import "./filters-bar.css";
-@import "./filters-options.css";
+@import "./filters-content.css";
 @import "./side-bar.css";
 
 html {
